Refactor, edit .gitignore
