+++ .gitignore
@@ -19,5 +19,10 @@
 *.tar.gz
 *.rar
 
+target/
+*~
+*.DS_Store
+/.idea/*
+
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
