+++ Procfile
@@ -1,2 +1,1 @@
-web: java $JAVA_OPTS -jar target/sparkjava-demo-01-1.0-jar-with-dependencies.jar 
-
+web: java $JAVA_OPTS -jar target/GoGaucho-1.0-jar-with-dependencies.jar
