+++ pom.xml
@@ -8,15 +8,15 @@
 
     <!-- project coordinates - values which uniquely identify this project -->
 
-  <groupId>edu.ucsb.cs56.pconrad</groupId>
-  <artifactId>sparkjava-demo-01</artifactId>
+  <groupId>edu.ucsb.cs56.gogaucho</groupId>
+  <artifactId>GoGaucho</artifactId>
   <version>1.0</version>
-  <url>https://ucsb-cs56-pconrad.github.io/sparkjava-01</url>
+  <url>https://github.com/a211212abc/ucsb-cs56-GoGaucho</url>
   
   <!-- configure what mvn package produces -->
   
   <packaging>jar</packaging>
-  <name>sparkjava-demo-01</name>
+  <name>GoGaucho</name>
   
   <!-- what version of Java should we use and what encoding for files? -->
   <properties>
@@ -93,7 +93,7 @@
             <configuration>
               <archive>
                 <manifest>
-		    <mainClass>edu.ucsb.cs56.pconrad.GoGaucho</mainClass>
+		    <mainClass>edu.ucsb.cs56.gogaucho.GoGaucho</mainClass>
                 </manifest>
               </archive>
               <descriptorRefs>
@@ -117,7 +117,7 @@
 	  </execution>
 	</executions>
 	<configuration>
-	  <mainClass>edu.ucsb.cs56.pconrad.GoGaucho</mainClass>
+	  <mainClass>edu.ucsb.cs56.gogaucho.GoGaucho</mainClass>
 	</configuration>
       </plugin>
 
@@ -133,7 +133,7 @@
           <processTypes>
             <!-- Tell Heroku how to launch your application -->
             <!-- You might have to remove the ./ in front   -->
-            <web>java $JAVA_OPTS -jar target/sparkjava-demo-01-1.0-jar-with-dependencies.jar</web>
+            <web>java $JAVA_OPTS -jar target/GoGaucho-1.0-jar-with-dependencies.jar</web>
 
           </processTypes>
 	</configuration>
