+++ src/main/java/edu/ucsb/cs56/gogaucho/GoGaucho.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.pconrad;
+package edu.ucsb.cs56.gogaucho;
 
 import static spark.Spark.port;
