+++ src/test/java/edu/ucsb/cs56/gogaucho/SparkDemo01Test.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.pconrad;
+package edu.ucsb.cs56.gogaucho;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
